requirements.txt
flask
numpy
pillow
tensorflow
scikit-learn

# venv\Scripts\activate
